requirements.txt
streamlit
tensorflow
numpy
pillow
python-multipart
fastapi
uvicorn
